Move client to resources package.
