+++ .github/tests-go-1.12/entrypoint.sh
@@ -2,6 +2,6 @@
 
 set -eux
 
-go test -v -coverprofile cover.out ./...
+go test -v -short -coverprofile cover.out ./...
 
 curl -s https://codecov.io/bash | bash -s -- -X fix -e GOLANG_VERSION
